+++ jobbergate-api/jobbergate_api/apps/job_scripts/models.py
@@ -1,8 +1,8 @@
 """
 Database model for the JobScript resource.
 """
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Table
-from sqlalchemy.sql import func
+from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, String, Table
+from sqlalchemy.sql import False_, func
 from sqlalchemy.sql.schema import Column
 
 from jobbergate_api.metadata import metadata
@@ -14,9 +14,10 @@ job_scripts_table = Table(
     Column("job_script_name", String, nullable=False, index=True),
     Column("job_script_description", String, default=""),
     Column("job_script_owner_email", String, nullable=False, index=True),
-    Column("application_id", ForeignKey("applications.id")),
+    Column("application_id", ForeignKey("applications.id"), nullable=True),
     Column("created_at", DateTime, nullable=False, default=func.now()),
     Column("updated_at", DateTime, nullable=False, default=func.now(), onupdate=func.now()),
+    Column("is_archived", Boolean, nullable=False, default=False_()),
 )
 
 searchable_fields = [