+++ jobbergate-cli/jobbergate_cli/subapps/job_scripts/app.py
@@ -9,10 +9,22 @@ import typer
 
 from jobbergate_cli.constants import SortOrder
 from jobbergate_cli.exceptions import Abort, handle_abort
-from jobbergate_cli.render import StyleMapper, render_json, render_list_results, render_single_result, terminal_message
+from jobbergate_cli.render import (
+    StyleMapper,
+    render_json,
+    render_list_results,
+    render_single_result,
+    terminal_confirm,
+    terminal_message,
+)
 from jobbergate_cli.requests import make_request
 from jobbergate_cli.schemas import JobbergateContext, JobScriptResponse, ListResponseEnvelope
-from jobbergate_cli.subapps.job_scripts.tools import create_job_script, download_job_script_files, fetch_job_script_data
+from jobbergate_cli.subapps.job_scripts.tools import (
+    change_archive_status,
+    create_job_script,
+    download_job_script_files,
+    fetch_job_script_data,
+)
 from jobbergate_cli.subapps.job_submissions.app import HIDDEN_FIELDS as JOB_SUBMISSION_HIDDEN_FIELDS
 from jobbergate_cli.subapps.job_submissions.tools import create_job_submission
 from jobbergate_cli.text_tools import dedent
@@ -24,6 +36,7 @@ HIDDEN_FIELDS = [
     "updated_at",
     "job_script_data_as_string",
     "job_script_files",
+    "is_archived",
 ]
 
 
@@ -41,6 +54,7 @@ app = typer.Typer(help="Commands to interact with job scripts")
 def list_all(
     ctx: typer.Context,
     show_all: bool = typer.Option(False, "--all", help="Show all job scripts, even the ones owned by others"),
+    include_archived: bool = typer.Option(False, help="Show archived job_scripts as well"),
     search: Optional[str] = typer.Option(None, help="Apply a search term to results"),
     sort_order: SortOrder = typer.Option(SortOrder.UNSORTED, help="Specify sort order"),
     sort_field: Optional[str] = typer.Option(None, help="The field by which results should be sorted"),
@@ -58,7 +72,7 @@ def list_all(
     assert jg_ctx is not None
     assert jg_ctx.client is not None
 
-    params: Dict[str, Any] = dict(all=show_all)
+    params: Dict[str, Any] = dict(all=show_all, include_archived=include_archived)
     if search is not None:
         params["search"] = search
     if sort_order is not SortOrder.UNSORTED:
@@ -86,7 +100,7 @@ def list_all(
         envelope,
         title="Job Scripts List",
         style_mapper=style_mapper,
-        hidden_fields=HIDDEN_FIELDS,
+        hidden_fields=HIDDEN_FIELDS + (["is_archived"] if include_archived else []),
     )
 
 
@@ -272,10 +286,33 @@ def delete(
         ...,
         help="The id of the job script to delete",
     ),
+    confirm: bool = typer.Option(
+        None,
+        "--confirm",
+        "-y",
+        help="If supplied, do not ask for confirmation; just delete.",
+    ),
 ):
     """
     Delete an existing job script.
     """
+    if not confirm and not terminal_confirm(
+        """
+        [yellow]Any files uploaded for this Job Script will be completely removed.
+        Any references to it in other items will also be removed.
+        """,
+        subject="Deleting is permanent!",
+        color="red",
+    ):
+        terminal_message(
+            """
+            No job_script was deleted.
+            """,
+            subject="Aborted",
+            color="yellow",
+        )
+        return
+
     jg_ctx: JobbergateContext = ctx.obj
 
     # Make static type checkers happy
@@ -293,6 +330,39 @@ def delete(
         "The job script was successfully deleted.",
         subject="Job script delete succeeded",
     )
+
+
+@app.command()
+@handle_abort
+def archive(
+    ctx: typer.Context,
+    id: int = typer.Option(
+        ...,
+        help="the specific id of the job_script to archive.",
+    ),
+):
+    """
+    Archive an existing job_script.
+
+    A job_script that is archived will not appear in lists by default. This is a way to
+    effectively hide a job_script from view.
+    """
+    change_archive_status(ctx.obj, id, True)
+
+
+@app.command()
+@handle_abort
+def restore(
+    ctx: typer.Context,
+    id: int = typer.Option(
+        ...,
+        help="the specific id of the job_script to restore from the archive.",
+    ),
+):
+    """
+    Restore an archived job_script.
+    """
+    change_archive_status(ctx.obj, id, False)
 
 
 @app.command()